requirements.txt
torch == 2.1.0
scipy
Pillow
numpy 
torchvision == 0.16
matplotlib
numpy-stl